refactor: consolidate developer tooling into debug domain
