+++ src/App.tsx
@@ -23,9 +23,9 @@ import { TestDeeplinkHandler } from '@/app/navigation/TestDeeplinkHandler';
 import { RainbowToastDisplay } from '@/components/rainbow-toast/RainbowToast';
 import { OfflineToast } from '@/components/toasts';
 import { reactNativeDisableYellowBox, showNetworkRequests, showNetworkResponses } from '@/config/debug';
-import monitorNetwork from '@/debugging/network';
 import { IS_DEV, IS_PROD, IS_STORE_INSTALL, IS_TEST } from '@/env';
 import { initializeRemoteConfig } from '@/features/config/stores/remoteConfig';
+import { monitorNetwork } from '@/features/debug/utils/network';
 import { configureDelegationSdk } from '@/features/delegation/configureClient';
 import RainbowContextWrapper from '@/helpers/RainbowContext';
 import { useApplicationSetup } from '@/hooks/useApplicationSetup';
