+++ src/features/debug/constants/diagnostics.ts
